# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateReturnMobileRequest:
    properties:
      channel_id:
        type: integer
      store_id:
        type: integer
      tracking:
        type: string
    required:
    - channel_id
    - store_id
    - tracking
    type: object
...
  controllers.MbRibbonFlowInfo:
    properties:
      created_at:
        type: string
      user:
        $ref: '#/definitions/controllers.RibbonUserFlowInfo'
    type: object
...
  controllers.QcRibbonFlowInfo:
    properties:
      created_at:
        type: string
      user:
        $ref: '#/definitions/controllers.RibbonUserFlowInfo'
    type: object
...
  controllers.ReturnMobilesListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utils.PaginationResponse'
      return_mobiles:
        items:
          $ref: '#/definitions/models.ReturnMobileResponse'
        type: array
    type: object
...
  controllers.RibbonFlowResponse:
    properties:
      mb_ribbon:
        $ref: '#/definitions/controllers.MbRibbonFlowInfo'
      order:
        $ref: '#/definitions/controllers.RibbonOrderFlowInfo'
      outbound:
        $ref: '#/definitions/controllers.RibbonOutboundFlowInfo'
      qc_ribbon:
        $ref: '#/definitions/controllers.QcRibbonFlowInfo'
      tracking:
        type: string
    type: object
  controllers.RibbonFlowsListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utils.PaginationResponse'
      ribbon_flows:
        items:
          $ref: '#/definitions/controllers.RibbonFlowResponse'
        type: array
    type: object
  controllers.RibbonOrderFlowInfo:
    properties:
      complained:
        type: boolean
      created_at:
        type: string
      order_ginee_id:
        type: string
      tracking:
        type: string
    type: object
  controllers.RibbonOutboundFlowInfo:
    properties:
      created_at:
        type: string
      expedition:
        type: string
      expedition_color:
        type: string
      user:
        $ref: '#/definitions/controllers.RibbonUserFlowInfo'
    type: object
  controllers.RibbonUserFlowInfo:
    properties:
      full_name:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  models.ChannelResponse:
    properties:
      code:
        example: SP
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ReturnMobileResponse:
    properties:
      channel:
        allOf:
        - $ref: '#/definitions/models.ChannelResponse'
        description: Related data
      channel_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      store:
        $ref: '#/definitions/models.StoreResponse'
      store_id:
        type: integer
      tracking:
        type: string
      updated_at:
        type: string
    type: object
...
  models.StoreResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: 192.168.31.136:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/mobile/returns:
    get:
      consumes:
      - application/json
      description: Get list of all return mobiles (logged-in users only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by return mobile tracking (partial match)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.ReturnMobilesListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all return mobiles
      tags:
      - return-mobiles
    post:
      consumes:
      - application/json
      description: Create a new return mobile (logged-in users only)
      parameters:
      - description: Create return mobile request
        in: body
        name: return_mobile
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateReturnMobileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ReturnMobileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new return mobile
      tags:
      - return-mobiles
  /api/mobile/returns/{id}:
    get:
      consumes:
      - application/json
      description: Get a return mobile by ID (logged-in users only)
      parameters:
      - description: Return Mobile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ReturnMobileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get a return mobile by ID
      tags:
      - return-mobiles
...
  /api/ribbon-flow:
    get:
      consumes:
      - application/json
      description: Get all ribbon flows with pagination and search, primary tracking
        from mb-ribbon (logged-in users only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Start date (YYYY-MM-DD or YYYY-M-D format)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD or YYYY-M-D format)
        in: query
        name: end_date
        type: string
      - description: Search by tracking number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.RibbonFlowsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all ribbon flows
      tags:
      - ribbon-flow
  /api/ribbon-flow/{tracking}:
    get:
      consumes:
      - application/json
      description: Get the complete flow tracking through ribbon process (mb-ribbon
        -> qc-ribbon -> outbound -> order) (logged-in users only)
      parameters:
      - description: Tracking number
        in: path
        name: tracking
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.RibbonFlowResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get ribbon flow tracking
      tags:
      - ribbon-flow
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abcdef12345'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"